Updates for Concede
Fix an issue and add backend validation to ensure concede can't be "hacked" into a post request in a game of the wrong variant type.

+++ gamemaster/game.php
@@ -1180,31 +1180,34 @@ class processGame extends Game
 	{
 		global $DB;
 
- 		// Unpause the game so that the processTime data isn't finalized as NULL
-		if( $this->processStatus == 'Paused' ) { $this->togglePause(); }
-
-		$this->archiveTerrStatus();
-		 
- 		if ( $this->phase == 'Diplomacy' and $this->turn > 0 )
+		// Ensure backend that a concede cannot be applied for the wrong variant type. 
+		if ( (empty(Config::$concedeVariants)) || (in_array($this->variantID, Config::$concedeVariants)) )
 		{
-			$DB->sql_put("INSERT INTO wD_MovesArchive
-				( gameID, turn, terrID, countryID, unitType, success, dislodged, type, toTerrID, fromTerrID, viaConvoy )
-				SELECT gameID, turn+1, terrID, countryID, unitType, success, dislodged, type, toTerrID, fromTerrID, viaConvoy
-				FROM wD_MovesArchive WHERE gameID = ".$this->id." AND turn = ".($this->turn-1));
+			// Unpause the game so that the processTime data isn't finalized as NULL
+			if( $this->processStatus == 'Paused' ) { $this->togglePause(); }
+
+			$this->archiveTerrStatus();
+			
+			if ( $this->phase == 'Diplomacy' and $this->turn > 0 )
+			{
+				$DB->sql_put("INSERT INTO wD_MovesArchive
+					( gameID, turn, terrID, countryID, unitType, success, dislodged, type, toTerrID, fromTerrID, viaConvoy )
+					SELECT gameID, turn+1, terrID, countryID, unitType, success, dislodged, type, toTerrID, fromTerrID, viaConvoy
+					FROM wD_MovesArchive WHERE gameID = ".$this->id." AND turn = ".($this->turn-1));
+			}
+			
+			$this->Members->setConcede();
+			foreach($this->Members->ByStatus['Playing'] as $Member)
+				$Winner = $Member;
+			$this->setWon($Winner);
+
+			$DB->sql_put("DELETE FROM wD_Orders WHERE gameID = ".$this->id);
+			$DB->sql_put("DELETE FROM wD_Units WHERE gameID = ".$this->id);
+			$DB->sql_put("DELETE FROM wD_TerrStatus WHERE gameID = ".$this->id);
+
+			Game::wipeCache($this->id,$this->turn);
 		}
-		
-		$this->Members->setConcede();
-		foreach($this->Members->ByStatus['Playing'] as $Member)
-			$Winner = $Member;
-		$this->setWon($Winner);
-
- 		$DB->sql_put("DELETE FROM wD_Orders WHERE gameID = ".$this->id);
-		$DB->sql_put("DELETE FROM wD_Units WHERE gameID = ".$this->id);
-		$DB->sql_put("DELETE FROM wD_TerrStatus WHERE gameID = ".$this->id);
-
- 		Game::wipeCache($this->id,$this->turn);
 	}
-	
 }
 
 ?>
